Remove the use of $releasever
The $releasever is perfectly fine for most installations. However, in the
instance that a system has a releasever set to something other than the
default, like 8.4, it causes issues as the EPEL repository structure
doesn't include these minor versions.

The solution here was a simple:

    sed 's/\$releasever/8/g' -i *.repo

Tested the change. Encountered a single issue with the
epel-playground-source repo, but it was unrelated and seen with the
unpatched revision.

Signed-off-by: Kyle Walker <kwalker@redhat.com>