Change modular repo names from DEPRECATED to RETIRED
As discussed at the EPEL Steering Committee meeting, change the warning
in the modular repo names from DEPRECATED to RETIRED to give a stronger
indication that they are no longer maintained.  It's uncessary to ship a
build just for this, so we'll defer it to go out in the next update of
the package.

https://pagure.io/epel/issue/198
https://meetbot.fedoraproject.org/fedora-meeting/2023-02-15/epel.2023-02-15-21.00.html
https://lists.fedoraproject.org/archives/list/epel-devel@lists.fedoraproject.org/thread/KA3JQUZW63A4V7OQ7ZVDFCM32EQUSOXH/